Chores : 팝업창 관련 모든 함수 파일로 분리하기 resolved :#91:
